+++ client/cmdlfnoralsy.c
@@ -108,6 +108,7 @@ int CmdNoralsyDemod(const char *Cmd) {
 		return 0;
 	}
 	setDemodBuf(DemodBuffer, 96, ans);
+	setGrid_Clock(32);
 	
 	//got a good demod
 	uint32_t raw1 = bytebits_to_byte(DemodBuffer, 32);